requirements.txt
Flask
Pillow
easyocr
navec
slovnet
torch
opencv-python
numpy
transformers
pytest
pytest-mock